+++ lib/src/models/text_formatter.dart
@@ -4,14 +4,13 @@ class FormattedTextFormatter {
   const FormattedTextFormatter({
     required this.patternChars,
     required this.style,
-  }) : pattern = r'\' +
-            patternChars +
-            '([^\\' +
-            patternChars +
-            ']+)\\' +
-            patternChars;
+  });
 
-  final String pattern;
   final String patternChars;
   final TextStyle style;
+
+  @override
+  String toString() {
+    return 'FormattedTextFormatter(patternChars: $patternChars, style: $style)';
+  }
 }